QEP DPDK and Linux kernel drivers release 1.0.0

+++ linux-kernel/driver/libqdma/libqdma_export.h
@@ -317,7 +317,7 @@ struct qdma_dev_conf {
 	 *  @msix_qvec_max: max. of vectors used for queues.
 	 *  libqdma update w/ actual #
 	 */
-	u8 msix_qvec_max;
+	u16 msix_qvec_max;
 	/** @uld: upper layer data, i.e. callback data */
 	unsigned long uld;
 	/** @qdma_drv_mode: qdma driver mode */
